+++ client/src/components/Header.js
@@ -1,9 +1,18 @@
-import React from 'react'; 
+import React, {useState} from 'react'; 
 
-function Header(props) {
+function Header() {
+    function Name() {
+        const [name, setName] = useState('Guest');
+        function login() {
+            setName("Username")
+        }
+    }
     return (
-        <h1>{props.title}</h1>
-    )
+        <div className="Homer">
+        <h1>{Name}</h1>
+        <p> Welcome to the Movie Site </p>
+        </div>
+    );
 }
 
 export default Header;
