+++ client/src/components/Login.js
@@ -1,4 +1,4 @@
-import React from "react"; 
+import React, {useState} from "react"; 
 import ProfileForm from "./ProfileForm";
 import { useHistory } from 'react-router-dom';
 
@@ -13,7 +13,7 @@ function Login() {
         history.push('/')
     }
 
-    return(ProfileForm)
+    return(<ProfileForm/>)
 }
